Link CMSIS-NN as static library
This links CMSIS-NN as a static library, rather than compiling it
together with in-tree source files. This allows CMSIS-NN to have a
separate optimization level and compiler flags as the rest of the code.

Compiling CMSIS-NN with `-O0` fails, which is not helpful when compiling
the remaining code for debugging.